+++ resources/wait_for_webapp.rb
@@ -20,7 +20,7 @@ actions :wait
 default_action :nothing
 
 attribute :host, kind_of: String, default: nil
-attribute :port, kind_of: Fixnum, default: 80
+attribute :port, kind_of: Integer, default: 80
 attribute :webapp, kind_of: String, name_attribute: true
-attribute :read_timeout, kind_of: Fixnum, default: 1800
-attribute :open_timeout, kind_of: Fixnum, default: 10
+attribute :read_timeout, kind_of: Integer, default: 1800
+attribute :open_timeout, kind_of: Integer, default: 10