+++ spec/install_kvm_spec.rb
@@ -13,9 +13,33 @@
 # limitations under the License.
 
 require 'spec_helper'
+require_relative 'support/matchers'
 
 describe 'abiquo::install_kvm' do
-  let(:chef_run) { ChefSpec::SoloRunner.new.converge(described_recipe) }
+  let(:chef_run) do
+    ChefSpec::SoloRunner.new do |node|
+      node.set['virtualization']['role'] = 'guest'
+    end.converge(described_recipe)
+  end
+  let(:c6_run) { ChefSpec::SoloRunner.new(platform: 'centos', version: '6.5') }
+
+  it 'locks the seabios version on CentOS 7' do
+    expect(chef_run).to create_yum_repository('kvm-common').with(
+      description: 'Backports for the seabios package',
+      baseurl: 'http://buildlogs.centos.org/centos/7/virt/x86_64/kvm-common',
+      includepkgs: 'seabios,seabios-bin,seavgabios-bin',
+      gpgcheck: false
+    )
+    expect(chef_run).to install_package('yum-plugin-versionlock')
+    expect(chef_run).to install_yum_package('seabios').with(version: '1.7.5-11.el7')
+    expect(chef_run).to lock_yum_package('seabios').with(version: '1.7.5-11.el7')
+
+    c6_run.converge(described_recipe)
+    expect(c6_run).to_not install_package('yum-plugin-versionlock')
+    expect(c6_run).to_not create_yum_repository('kvm-common')
+    expect(c6_run).to_not install_yum_package('seabios')
+    expect(c6_run).to_not lock_yum_package('seabios')
+  end
 
   %w(qemu-kvm abiquo-aim abiquo-sosreport-plugins).each do |pkg|
     it "installs the #{pkg} package" do