+++ spec/setup_frontend_spec.rb
@@ -18,13 +18,13 @@ require_relative 'support/stubs'
 describe 'abiquo::setup_frontend' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new do |node|
-      node.set['abiquo']['certificate']['common_name'] = 'test.local'
+      node.set['abiquo']['certificate']['common_name'] = 'fauxhai.local'
     end.converge('apache2::default', 'abiquo::install_websockify', described_recipe, 'abiquo::service')
   end
-  let(:cn) { 'test.local' }
+  let(:cn) { 'fauxhai.local' }
 
   before do
-    stub_certificate_files('/etc/pki/abiquo/test.local.crt', '/etc/pki/abiquo/test.local.key')
+    stub_certificate_files('/etc/pki/abiquo/fauxhai.local.crt', '/etc/pki/abiquo/fauxhai.local.key')
     stub_command('/usr/sbin/httpd -t').and_return(true)
     stub_command("/usr/bin/test -f /etc/pki/abiquo/#{cn}.crt").and_return(true)
     stub_command('/usr/bin/mysql kinton -e \'SELECT 1\'').and_return(true)