+++ spec/setup_server_spec.rb
@@ -19,7 +19,7 @@ require_relative 'support/stubs'
 describe 'abiquo::setup_server' do
   let(:chef_run) do
     ChefSpec::SoloRunner.new do |node|
-      node.set['abiquo']['certificate']['common_name'] = 'test.local'
+      node.set['abiquo']['certificate']['common_name'] = 'fauxhai.local'
     end
   end
 
@@ -27,7 +27,7 @@ describe 'abiquo::setup_server' do
     stub_queries
     stub_command('/usr/sbin/httpd -t').and_return(true)
     stub_command("/usr/bin/mysql kinton -e 'SELECT 1'").and_return(false)
-    stub_certificate_files('/etc/pki/abiquo/test.local.crt', '/etc/pki/abiquo/test.local.key')
+    stub_certificate_files('/etc/pki/abiquo/fauxhai.local.crt', '/etc/pki/abiquo/fauxhai.local.key')
   end
 
   it 'includes the service recipe' do